requirements/dev.txt
-r common.txt
django-debug-toolbar=4.4.6
django-extensions==3.2.3
ipython==8.29.0